refactor(file): refactor code to implement single responsibility principle
