+++ src/index.html
@@ -10,7 +10,7 @@
 
     <script type="text/javascript" src="libs/jquery-1.12.4.min.js"></script>
     <script type="text/javascript" src="libs/materialize.min.js"></script>
-    
+
     <script type="text/javascript" src="js/inverted-index.js"></script>
     <script type="text/javascript" src="js/index.js"></script>
     <script type="text/javascript" src="js/helpers.js"></script>
